# chore(dev): reload in dev now takes ~500ms