--- a/.gitignore
+++ b/.gitignore
@@ -8,3 +8,4 @@ npm-debug.log
 .idea/
 *.iml
 preview.zip
+.happypack/