--- a/docgen/layouts/archetypes/content-with-menu.pug
+++ b/docgen/layouts/archetypes/content-with-menu.pug
@@ -11,4 +11,5 @@ body.documentation
       .documentation-container
         block content
   include ../common/footer.pug
+  script(src=webpack.assets['js/common.js'])
   script(src=webpack.assets['js/main.js'])