--- a/docgen/layouts/archetypes/single-column-formatted.pug
+++ b/docgen/layouts/archetypes/single-column-formatted.pug
@@ -7,4 +7,5 @@ body.documentation
     .container
       block content
   include ../common/footer.pug
+  script(src=webpack.assets['js/common.js'])
   script(src=webpack.assets['js/main.js'])