--- a/docgen/layouts/archetypes/single-column.pug
+++ b/docgen/layouts/archetypes/single-column.pug
@@ -6,4 +6,5 @@ body
   div
     block content
   include ../common/footer.pug
+  script(src=webpack.assets['js/common.js'])
   script(src=webpack.assets['js/main.js'])