--- a/docgen/layouts/example.pug
+++ b/docgen/layouts/example.pug
@@ -15,6 +15,7 @@ html
   body.example
     include ./common/header.pug
     .content!=contents
+    script(src=webpack.assets['js/common.js'])
     script(src=webpack.assets['js/main.js'])
     script(src=webpack.assets[`examples/${name}/index.js`])
     script(src="https://cdn.jsdelivr.net/g/docsearch.js@2")