--- a/packages/react-instantsearch-theme-algolia/package.json
+++ b/packages/react-instantsearch-theme-algolia/package.json
@@ -13,10 +13,10 @@
     "build-and-publish": "./scripts/build-and-publish.sh"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7.3",
+    "autoprefixer": "^6.7.4",
     "cssnano": "^3.10.0",
     "node-sass": "^4.5.0",
-    "postcss": "^5.2.13",
+    "postcss": "^5.2.14",
     "postcss-scss": "^0.4.1"
   },
   "author": {