--- a/packages/react-instantsearch-theme-algolia/postcss.config.js
+++ b/packages/react-instantsearch-theme-algolia/postcss.config.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/no-commonjs */
+// This cannot use ES6 exports, it will be read by something that only
+// reads commonjs exports
+module.exports = {
+  plugins: [
+    require('autoprefixer'),
+  ],
+};