--- a/packages/react-instantsearch/examples/autocomplete/package.json
+++ b/packages/react-instantsearch/examples/autocomplete/package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "lodash": "^4.16.6",
     "react": "^15.3.2",
-    "react-autosuggest": "^8.0.0",
+    "react-autosuggest": "^8.0.1",
     "react-dom": "^15.3.2",
     "react-instantsearch": "latest",
     "react-instantsearch-theme-algolia": "latest"