--- a/packages/react-instantsearch/webpack.config.babel.js
+++ b/packages/react-instantsearch/webpack.config.babel.js
@@ -16,9 +16,9 @@ export default {
     libraryTarget: 'umd',
   },
   module: {
-    loaders: [
+    rules: [
       {
-        test: /\.js$/, exclude: /node_modules/, loader: 'babel',
+        test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader',
       },
     ],
   },
@@ -36,10 +36,8 @@ export default {
     },
   },
   plugins: [
-    new webpack.DefinePlugin({
-      'process.env': {
-        NODE_ENV: JSON.stringify(process.env.NODE_ENV),
-      },
+    new webpack.EnvironmentPlugin({
+      NODE_ENV: 'development', // use 'development' unless process.env.NODE_ENV is defined
     }),
   ],
 };