--- a/scripts/boot.sh
+++ b/scripts/boot.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
 # Use this script to install dependencies of every packages
-
+rm -rf node_modules packages/react-instantsearch/node_modules packages/react-instantsearch-theme-algolia/node_modules &&
 yarn &&
-(cd packages/react-instantsearch && yarn) &&
-(cd packages/react-instantsearch-theme-algolia && yarn)
+(cd node_modules/react && yarn link) &&
+(cd node_modules/react-dom && yarn link) &&
+(cd packages/react-instantsearch && yarn && yarn link && yarn link react && yarn link react-dom) &&
+(cd packages/react-instantsearch-theme-algolia && yarn && yarn link && yarn link react && yarn link react-dom) &&
+yarn link react-instantsearch &&
+yarn link react-instantsearch-theme-algolia