… package.json
… packages/react-instantsearch-theme-algolia/yarn.lock
… packages/react-instantsearch/yarn.lock
… yarn.lock