Before this commit, the `yarn start` workflow took 10s when a file was CHANGED (not even talking about the 40s boot time), now it takes 500ms in average for a reload. Which is good.

Optimization was done via:
- webpack 2 switch
- dev env uses simpler sourcemap strategy
- use happypack (https://github.com/amireh/happypack)
- Create a common chunks plugin (this reduces the size of downloaded
JavaScript files by a big factor).

Flow example: load, one save, another save.

Before:
![image](https://cloud.githubusercontent.com/assets/123822/23194452/004d18e2-f863-11e6-8ed4-18c32883c1af.png)

Startup took ~50s. Saves took ~10s each.

After:

![image](https://cloud.githubusercontent.com/assets/123822/23194507/59bd95e6-f863-11e6-8b7e-7fcd263d647d.png)

Startup took ~30s. Saves took ~500ms (Ok FIRST save took 2s, but trust me, every other took 500ms)